requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colored-traceback==0.3.0
commonmark==0.9.1
cryptography==36.0.2
cssselect==1.1.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.11.3
intervaltree==3.1.0
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.10.3
parse==1.19.0
plumbum==1.7.2
psutil==5.9.0
pwntools==4.7.0
pycparser==2.21
pyee==8.2.2
pyelftools==0.28
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
requests-html==0.10.0
rich==12.2.0
ROPGadget==6.7
rpyc==5.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
unicorn==2.0.0rc7
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
zipp==3.8.0
